backend/requirements.txt
django
djangorestframework
djoser
corsheaders
django-cors-headers
